DABOM-305 refactor: 스타일 가이드 준수 (from 팩토리, readOnly, 패키지 분리)
- TokenRefreshResponse를 global/auth → global/api/response 패키지로 이동
- TokenRefreshResponse.from(TokenRefreshResult) 정적 팩토리 메서드 추가
- AdminController, CustomerController에서 from() 팩토리 메서드 사용
- AdminServiceImpl, CustomerServiceImpl의 refreshToken()에 @Transactional(readOnly = true) 추가
