+++ src/main/java/com/project/domain/customer/controller/CustomerController.java
@@ -14,7 +14,7 @@ import com.project.domain.customer.dto.response.SignInResponse;
 import com.project.domain.customer.dto.response.SignUpResponse;
 import com.project.domain.customer.service.CustomerService;
 import com.project.global.api.response.ApiResponse;
-import com.project.global.auth.TokenRefreshResponse;
+import com.project.global.api.response.TokenRefreshResponse;
 import com.project.global.auth.TokenRefreshResult;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -49,9 +49,7 @@ public class CustomerController {
     public ApiResponse<TokenRefreshResponse> refreshToken(
             @Valid @RequestBody CustomerRefreshRequest request) {
         TokenRefreshResult result = customerService.refreshToken(request.refreshToken());
-        return ApiResponse.success(
-                new TokenRefreshResponse(
-                        result.accessToken(), result.refreshToken(), result.expiresIn()));
+        return ApiResponse.success(TokenRefreshResponse.from(result));
     }
 
     @PostMapping("/logout")
