--- src/main/java/com/project/global/auth/TokenRefreshResponse.java
@@ -1,3 +1,0 @@
-package com.project.global.auth;
-
-public record TokenRefreshResponse(String accessToken, String refreshToken, long expiresIn) {}
